First beginnings of returning music metadata via a Nan module.
Currently images are being recoginized and opened like an audio track.  The musicData module should return data about if the file is an image or what.

Next steps is to upgrade musicData to be able to return attachment binary data (album artwork) and possibly codec info for future reads to this file.